CS10 -Reyhan Batara Revisi

+++ frontend_ReyhanBatara/core/src/main/java/com/Reyhan/frontend/strategies/DifficultyStrategy.java
@@ -4,7 +4,7 @@ import java.util.Map;
 
 public interface DifficultyStrategy {
     float getSpawnInterval();
-    int getDensity();
+    float getDensity();
     float getMinGap();
     Map<String, Integer> getObstacleWeights();
 }
